1d->3d mappings
corrected some errors in synapse mapping, but something still seems to be fishy in the network case: active synapse locations in 3d do not all correspond to active potential areas in 1d.
Something must also be wrong with the potential mapping. 1d and 3d representations do not show the same behavior.