Implemented a voltage-explicit exact gating mode for HH defect assembling
This mode can be used if (and only if) the time discretization scheme for the gating parameters is explicit Euler.
It will solve the ODEs governing the gating variables exactly in every time step, using the voltage from the last time step.
Activate this mode in the lua script by:
    hodgkinHuxleyChannels:use_exact_gating_mode(dt_dimless)
where dt_dimless is the step size that is used in each time step in the explicit Euler.