Adding custom tetrahedralize method and moving MergeGrids to UG4.
A custom tetrahedralize method which operates on a selection of the grid
by using a Selector was introduced to fill the soma and ER parts of the
geometry. MergeGrids function was moved to UG4's lib_grid file utils.
Note that now now grid copying is necessary, tetrahedralize will work on
one grid. This will circumvent the previously encountered problem that
grid attachments vanished during CopyGrid function in lib_grid file
utils, which is still not resolved.