Correcting MitigateRootBranchingNeurites.
Direct branching at soma is not allowed for now. Move the branches one
point forward on the root/main branch. This will introduce branches with
more than 3. Inserting an additional point between soma and branch will
not work since smoothing will remove the generated edges in the subsequent
process. This needs to be improved somehow.