Adding PathLength1D.
The method can be used to measure the distance between two (measuring)
subsets, e.g. meas1 and meas2. The method requires a 3D mesh and
optionally the corresponding 1D mesh to calculate the path length with
the Dijkstra algorithm (positive edge weights) using a priority queue.
Added a very basic unit test as well to test the functionality.